Move lp, long, and short  portfolio tables under portfolio/
