+++ apps/hyperdrive-trading/src/ui/hyperdrive/longs/CloseLongForm/CloseLongForm.tsx
@@ -17,8 +17,8 @@ import { useNumericInput } from "src/ui/base/hooks/useNumericInput";
 import { ConnectWalletButton } from "src/ui/compliance/ConnectWallet";
 import { useCloseLong } from "src/ui/hyperdrive/longs/hooks/useCloseLong";
 import { usePreviewCloseLong } from "src/ui/hyperdrive/longs/hooks/usePreviewCloseLong";
-import { StatusCell } from "src/ui/hyperdrive/longs/OpenLongsTable/StatusCell";
 import { TransactionView } from "src/ui/hyperdrive/TransactionView";
+import { StatusCell } from "src/ui/portfolio/longs/OpenLongsTable/StatusCell";
 import { useTokenBalance } from "src/ui/token/hooks/useTokenBalance";
 import { useTokenFiatPrice } from "src/ui/token/hooks/useTokenFiatPrice";
 import { TokenInput } from "src/ui/token/TokenInput";
