+++ apps/hyperdrive-trading/src/ui/portfolio/longs/OpenLongsTable/CurrentValueCell.tsx
