+++ apps/hyperdrive-trading/src/ui/portfolio/longs/OpenLongsTable/TotalOpenLongsValue.tsx
