+++ apps/hyperdrive-trading/src/ui/portfolio/lp/LpAndWithdrawalSharesTable/SizeAndPoolShareCell.tsx
