+++ apps/hyperdrive-trading/src/ui/portfolio/shorts/OpenShortsTable/AccruedYieldCell.tsx
