+++ apps/hyperdrive-trading/src/ui/portfolio/shorts/OpenShortsTable/CurrentShortsValueCell.tsx
