+++ apps/hyperdrive-trading/src/ui/portfolio/shorts/OpenShortsTable/CurrentValueCell.tsx
