+++ apps/hyperdrive-trading/src/ui/portfolio/shorts/OpenShortsTable/ShortRateAndSizeCell.tsx
