+++ apps/hyperdrive-trading/src/ui/portfolio/shorts/OpenShortsTable/TotalOpenShortsValue.tsx
